Fix recaptcha loading in Plone 4 modal
The way plone.app.jquerytools handles the modal, jquery removes the
script tags from the html loaded in the modal. It looks like javascript
code inside the script tag is still executed, but the tag itself is
removed. As the recaptcha javascript is in the src attribute, it is not
executed once the tag is removed.

So we added a javascript that dynamically add the script tag with src
from recaptcha, If the request is ajax.